Finished all functions at the field seminar. NOT tested

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: dnovak <dnovak@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/20 16:25:59 by dnovak            #+#    #+#             */
-/*   Updated: 2024/05/20 16:43:40 by dnovak           ###   ########.fr       */
+/*   Updated: 2024/05/21 00:18:30 by dnovak           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,9 +21,9 @@ static int	ft_isspace(int c)
 
 int	ft_atoi(const char *nptr)
 {
-	int	i;
-	int	sign;
-	int	num;
+	size_t	i;
+	int		sign;
+	int		num;
 
 	i = 0;
 	while (ft_isspace(*(nptr + i)))
